feat: pesos de los producto, sweetalert2 y validación de producto

+++ resources/views/admin/gathering/create.blade.php
@@ -69,15 +69,15 @@
                 <div class="row">
                     <div class="col-md-3">
                         <label for="carton_weight" class="form-label">{{ __('Carton Weight') }}</label>
-                        <input type="text" class="form-control" name="" id="carton_weight" disabled>
+                        <input type="number" class="form-control" name="carton_weight" id="carton_weight" disabled>
                     </div>
                     <div class="col-md-3">
                         <label for="plastic_weight" class="form-label">{{ __('Plastic Weight') }}</label>
-                        <input type="text" class="form-control" name="" id="plastic_weight" disabled>
+                        <input type="text" class="form-control" name="plastic_weight" id="plastic_weight" disabled>
                     </div>
                     <div class="col-md-3">
                         <label for="paper_weight" class="form-label">{{ __('Paper Weight') }}</label>
-                        <input type="text" class="form-control" name="" id="paper_weight" disabled>
+                        <input type="text" class="form-control" name="paper_weight" id="paper_weight" disabled>
                     </div>
                     <div class="col-md-3">
                         <label for="overall_weight" class="form-label">{{ __('Overall Weight') }}</label>
@@ -148,6 +148,7 @@
 @endpush
 
 @push('scripts')
+    <script src="{{ asset('plugins/sweetalert2/sweetalert2.all.min.js') }}"></script>
     <script>
         $('select[name=product_id]').on('change',function() {
             $('select[name=product_type_id]').html('<option value="" selected disabled>{{ __("Select a type of product") }}</option>');
@@ -160,8 +161,12 @@
         });
 
         var cont = 1;
-        overall_weight = 0;
-        product_weight = [];
+        product_id_array = [];      // Para guardar los id de cada producto
+        product_weight = [];        // Para guardar los pesos de cada producto
+        total_carton_weight = 0;    // Para guardar peso total de carton
+        total_plastic_weight = 0;   // Para guardar peso total de plastico
+        total_paper_weight = 0;     // Para guardar peso total de papel
+        overall_weight = 0;         // Para guardar peso total general
 
         $('#addProduct').click(function () { 
             productData = document.getElementById('product_id').value.split('_');
@@ -172,25 +177,31 @@
             productType = $('#product_type_id option:selected').text();
             packages = $('#packages').val();
             weight = $('#weight').val();
-            // var cont = 1;
-            // overall_weight = 0;
-            // product_weight = [];
-            product_weight[cont] = parseInt(weight);
-            overall_weight = overall_weight + product_weight[cont];
-            console.log(overall_weight);
-            // Validar despues, que no manden campos vacios
-            var row = '<tr id="row'+cont+'">'+
+            if (product_id != "select" && product_type_id != "" && packages != "" && weight != "") {
+                product_id_array[cont] = parseInt(product_id);
+                product_weight[cont] = parseInt(weight);
+                overall_weight = overall_weight + product_weight[cont];
+                var row = '<tr id="row'+cont+'">'+
                     '<td><input type="hidden" name="product_id[]" value="'+product_id+'">'+product+'</td>'+
                     '<td><input type="hidden" name="product_type_id[]" value="'+product_type_id+'">'+productType+'</td>'+
                     '<td><input type="hidden" name="packages[]" value="'+packages+'">'+packages+'</td>'+
                     '<td><input type="hidden" name="weight[]" value="'+weight+'">'+weight+'</td>'+
                     '<td><button type="button" class="btn btn-danger btn-sm" onclick="delete_row('+cont+');"><i class="fa-solid fa-xmark"></i></button></td>'+
                 '</tr>';
-            cont++;
-
-            $('#overall_weight').val(overall_weight);
-            $('#details').append(row);
-            clean();
+                cont++;
+                clean();
+                totals();
+                $('#carton_weight').val(total_carton_weight);
+                $('#plastic_weight').val(total_plastic_weight);
+                $('#paper_weight').val(total_paper_weight);
+                $('#overall_weight').val(overall_weight);
+                $('#details').append(row);
+            } else {
+                Swal.fire({
+                    icon: 'error',
+                    text: 'Rellene todos los campos del producto.',
+                })
+            }
         });
 
         function clean() {
@@ -201,8 +212,31 @@
             $("#product_type_id").html('<option value="">{{ __("Select a type of product") }}</option>');
         }
 
+        function totals() {
+            if (product_id == 1) {
+                product_weight[cont] = parseInt(weight);
+                total_carton_weight = total_carton_weight + product_weight[cont];
+            } else if (product_id == 2) {
+                product_weight[cont] = parseInt(weight);
+                total_plastic_weight = total_plastic_weight + product_weight[cont];
+            } else if (product_id == 3) {
+                product_weight[cont] = parseInt(weight);
+                total_paper_weight = total_paper_weight + product_weight[cont];
+            }
+        }
+
         function delete_row(index) {
+            if (product_id_array[index] == 1) {
+                total_carton_weight = total_carton_weight - product_weight[index];
+            } else if (product_id_array[index] == 2) {
+                total_plastic_weight = total_plastic_weight - product_weight[index];
+            } else if (product_id_array[index] == 3){
+                total_paper_weight = total_paper_weight - product_weight[index];
+            }
             overall_weight = overall_weight - product_weight[index];
+            $('#carton_weight').val(total_carton_weight);
+            $('#plastic_weight').val(total_plastic_weight);
+            $('#paper_weight').val(total_paper_weight);
             $('#overall_weight').val(overall_weight);
             $("#row"+index).remove();
         }
